requirements.txt
attrs==24.3.0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.221 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-aws-lambda-python-alpha==2.177.0a0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-cloud-assembly-schema==39.2.7 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-lib==2.177.0 ; python_version >= "3.12" and python_version < "4.0"
boto3==1.36.7 ; python_version >= "3.12" and python_version < "4.0"
botocore==1.36.7 ; python_version >= "3.12" and python_version < "4.0"
cattrs==24.1.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
constructs==10.4.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-resources==6.5.2 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
jsii==1.106.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.4 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
s3transfer==0.11.2 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
